Moved helper functions into utils.py
